+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 2.9.10
 - Wrap phone inside tel dial link in GetStuList.fnc.php
 - User Fields: search Email Address & Phone in CustomFields.fnc.php, GetStaffList.fnc.php, Search.fnc.php & Preferences.php
 - Add makeEmail() function in GetStuList.fnc.php
+- Help texts updates in Help_en.php & Help_es.php
 
 Changes in 2.9.9
 ----------------